+++ examples/resources/voipms_did/resource.tf
@@ -22,10 +22,10 @@ resource "voipms_did" "home" {
   routing              = "account:${data.voipms_subaccount.gateway.account}"
   pop_hostname         = local.my_pops.ny7
   dialtime             = 30
-  voicemail            = data.voipms_voicemail.johns_voicemail.mailbox
-  failover_busy        = "vm:${data.voipms_voicemail.johns_voicemail.mailbox}"
-  failover_noanswer    = "vm:${data.voipms_voicemail.johns_voicemail.mailbox}"
-  failover_unreachable = "vm:${data.voipms_voicemail.johns_voicemail.mailbox}"
+  voicemail_name       = data.voipms_voicemail.johns_voicemail.name
+  failover_busy        = "vm:${data.voipms_voicemail.johns_voicemail.name}"
+  failover_noanswer    = "vm:${data.voipms_voicemail.johns_voicemail.name}"
+  failover_unreachable = "fwd:${data.voipms_forwarding.mobile.description}"
 
   sms_enabled     = true
   webhook         = "https://example.lambda-url.us-east-1.on.aws/"